Move machine client into pkg/client/machine
