+++ hack/update-codegen.sh
@@ -21,10 +21,10 @@ set -o pipefail
 export GOPATH=$(go env GOPATH)
 
 echo Removing old client
-rm -rf "pkg/crd/client"
+rm -rf "pkg/client"
 SCRIPT_ROOT=$(dirname "${BASH_SOURCE}")
 echo $SCRIPT_ROOT
 ./vendor/k8s.io/code-generator/generate-groups.sh all \
-    github.com/kubermatic/machine-controller/pkg/client github.com/kubermatic/machine-controller/pkg \
+    github.com/kubermatic/machine-controller/pkg/client/machine github.com/kubermatic/machine-controller/pkg \
     machines:v1alpha1 \
     --go-header-file=${SCRIPT_ROOT}/header.txt
