+++ pkg/client/machine/clientset/versioned/clientset.go
@@ -20,7 +20,7 @@ package versioned
 
 import (
 	glog "github.com/golang/glog"
-	machinev1alpha1 "github.com/kubermatic/machine-controller/pkg/client/clientset/versioned/typed/machines/v1alpha1"
+	machinev1alpha1 "github.com/kubermatic/machine-controller/pkg/client/machine/clientset/versioned/typed/machines/v1alpha1"
 	discovery "k8s.io/client-go/discovery"
 	rest "k8s.io/client-go/rest"
 	flowcontrol "k8s.io/client-go/util/flowcontrol"
