+++ pkg/client/machine/clientset/versioned/fake/clientset_generated.go
@@ -19,9 +19,9 @@ limitations under the License.
 package fake
 
 import (
-	clientset "github.com/kubermatic/machine-controller/pkg/client/clientset/versioned"
-	machinev1alpha1 "github.com/kubermatic/machine-controller/pkg/client/clientset/versioned/typed/machines/v1alpha1"
-	fakemachinev1alpha1 "github.com/kubermatic/machine-controller/pkg/client/clientset/versioned/typed/machines/v1alpha1/fake"
+	clientset "github.com/kubermatic/machine-controller/pkg/client/machine/clientset/versioned"
+	machinev1alpha1 "github.com/kubermatic/machine-controller/pkg/client/machine/clientset/versioned/typed/machines/v1alpha1"
+	fakemachinev1alpha1 "github.com/kubermatic/machine-controller/pkg/client/machine/clientset/versioned/typed/machines/v1alpha1/fake"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/discovery"
