+++ pkg/client/machine/clientset/versioned/fake/register.go
