+++ pkg/client/machine/clientset/versioned/scheme/doc.go
