+++ pkg/client/machine/clientset/versioned/scheme/register.go
