+++ pkg/client/machine/clientset/versioned/typed/machines/v1alpha1/fake/fake_machines_client.go
@@ -19,7 +19,7 @@ limitations under the License.
 package fake
 
 import (
-	v1alpha1 "github.com/kubermatic/machine-controller/pkg/client/clientset/versioned/typed/machines/v1alpha1"
+	v1alpha1 "github.com/kubermatic/machine-controller/pkg/client/machine/clientset/versioned/typed/machines/v1alpha1"
 	rest "k8s.io/client-go/rest"
 	testing "k8s.io/client-go/testing"
 )
